WIP: drm/meson/meson_vclk: fix VIC alternate timings
4K VP9 media with 59.94 fractional refresh rate refuses to play with
the screen going blank (no sync) and the following error reported in
system logs:

[   89.610280] Fatal Error, invalid HDMI vclk freq 593406

Modetest shows the following:

3840x2160 59.94 3840 4016 4104 4400 2160 2168 2178 2250 593407 flags: phsync, pvsync; type: driver
calculated value -----------------------------------------^

Tweaking DIV_ROUND_CLOSEST to use 1002 in calculations allows 593406
to match and 4K 59.94 media now plays correctly.

Unfortunately this is not a proper fix and it's not guaranteed to
work for everyone. It points and issue with clocks/precision somewhere.

Fixes: ff217bc710e0 ("drm/meson: Add support for VIC alternate timings")
Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>